% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dChern.R
\name{dChern}
\alias{dChern}
\title{Density function of Chernoff's distribution}
\usage{
dChern(x)
}
\arguments{
\item{x}{evaluation point of the density.}
}
\value{
The function returns Chernoff's density evaluated at x.
}
\description{
Computes the density of Chernoff's distribution.
}
\examples{
dChern(0)
}
