% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pChern.R
\name{pChern}
\alias{pChern}
\title{Cumulative distribution function of Chernoff's distribution}
\usage{
pChern(q)
}
\arguments{
\item{q}{evaluation point of the distribution function.}
}
\value{
The function returns Chernoff's distribution function evaluated at q.
}
\description{
Computes the CDF of Chernoff's distribution.
}
\examples{
pChern(0)
}
