% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qChern.R
\name{qChern}
\alias{qChern}
\title{Quantile function of Chernoff's distribution}
\usage{
qChern(p)
}
\arguments{
\item{p}{evaluation point of the quantile function.}
}
\value{
The function returns Chernoff's quantile function evaluated at p.
}
\description{
Computes the quantiles of Chernoff's distribution.
}
\examples{
qChern(0.5)
}
